requirements.txt
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
eventlet==0.35.2
Flask==3.0.2
Flask-SocketIO==5.3.6
googlesearch-python==1.2.3
greenlet==3.0.3
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-engineio==4.9.0
python-socketio==5.11.1
requests==2.31.0
simple-websocket==1.0.0
soupsieve==2.5
urllib3==2.2.1
Werkzeug==3.0.1
wsproto==1.2.0

